fix: horizontal seam detection was unable to traverse across the entire width of the picture. See details for more information.
Background:
Horizontal seam detection is computing the least-energy pixel path across the horizontal width of the picture. Using the
dual-gradient energy function - the differences of adjacent pixel RGB values - Dijkstra's algorithim is implemented to compute
the path of minimum weight (the sum of all pixel energy on a given path). The weights of the left-most column are initiliazed to zero
and computes a minimum path to the right-most column of the picture. Accordingly, the right-most column should all contain finite values
however this was not the case, meaning the algorithim was unable to travel to the right-most column.

This was caused by incorrect index insertion during initialization, the index values being inserted to the priority queue were only the
row index, however not the actual row X col index. Fixed by simply correcting the initilization procedure to insert the row X col index.