Performance & Refactor: (Loaded Commit) Increase performance by up to 72% - See details for more information.
Changes:
-Replace Dijkstra's algorithim with Bellman Ford's algorithim. Increased performance dramatically by 20%. Although
in theory they should have very similar run-time I believe this difference is caused by the replacement of the PQ
to a normal queue in Bellman Ford's algorithim. I.e. This difference was most notable in large pictures of upwards
(1200px x 800px) therefore nearly a million pixels. Therefore Bellman's Ford algorithim would only require O(N) time
to check  boolean array if the pixel is already on the queue, while for the Dijkstra implementation I implemented, would
require O(N log N) time

-Replace the array implementation to track the energy of pixels with an ArrayList implementation. Dramatically increased
performance by 50%. I highly suspected this as I was previously using a single array which would require
near exponential run-time O(Width x Height) to copy & remove eligible pixels to a new array for each seam removal cycle.
The ArrayList implementation allowed for linear runtime for each seam removal, as the nature of the data structure
automatically shifts and resizes with each .delete() command. Given most common seam removal is for vertical seam removal
the ArrayList is implemented in the way to maximize performance for vertical seam removal. Horizontal seam removal
requires an additional linear runtime dependent on the difference between the row of the pixel being removed
and the height of the picture.

-Replace single array implementation of a few data structures with a double indexed array. Saved memory proportional to
the number of pixels within the picture, in addition to saving near exponential runtime relative to the original
method of implmentation.

-Refactored code and added comments